Update Dependencies Versions
Changes in the toml and lock file.

These are the most recent dependencies that were compatiable with our current version of the codebase.

I also added the cuda 11_8 for windows exe file as one needs to install the cudatoolkit in order to use cuda.

Also just ran codebase to confirm that the pyproject dependencies are working.